Move cycle property tests out of IR module
Relocate the `canonicalize_cycle` property tests into a test-only
include file so `src/ir/cycle.rs` stays below the project file-size
limit without changing production cycle-detection behaviour.
